fts: extract reader/phrase.rs (PhraseMember, PhraseCursor, AnyCursor)
Phase 1b of the read-path split. Move exact-phrase matching — the
positional PhraseCursor over PhraseMember atoms plus the AnyCursor enum
that unifies term and phrase atoms for the walks — out of core into
reader/phrase.rs. Reads the cursor layer through the pub(super) surface
from Phase 1a; its own types/fields/methods are pub(super) so core's
atom walks keep using them. Widened NormTable::{get,empty} to pub(super)
for the phrase scorer. Pure move — FTS oracle/count/phrase suites green.
